requirements.txt
python-telegram-bot==20.7
aiohttp
requests
deep-translator
feedparser
openai>=1.0.0